+++ Cargo.toml
@@ -17,6 +17,7 @@ cfg-if = "1.0.3"
 errno = "0.3.0"
 fish-build-helper = { path = "crates/build-helper" }
 fish-build-man-pages = { path = "crates/build-man-pages" }
+fish-gettext = { path = "crates/gettext" }
 fish-gettext-extraction = { path = "crates/gettext-extraction" }
 fish-gettext-maps = { path = "crates/gettext-maps" }
 fish-gettext-mo-file-parser = { path = "crates/gettext-mo-file-parser" }
@@ -87,8 +88,8 @@ cfg-if.workspace = true
 errno.workspace = true
 fish-build-helper.workspace = true
 fish-build-man-pages = { workspace = true, optional = true }
+fish-gettext = { workspace = true, optional = true }
 fish-gettext-extraction = { workspace = true, optional = true }
-fish-gettext-maps = { workspace = true, optional = true }
 fish-printf.workspace = true
 fish-tempfile.workspace = true
 libc.workspace = true
@@ -98,7 +99,6 @@ nix.workspace = true
 num-traits.workspace = true
 once_cell.workspace = true
 pcre2.workspace = true
-phf = { workspace = true, optional = true }
 rand.workspace = true
 terminfo.workspace = true
 xterm-color.workspace = true
@@ -156,7 +156,7 @@ benchmark = []
 embed-manpages = ["dep:fish-build-man-pages"]
 # Enable gettext localization at runtime. Requires the `msgfmt` tool to generate catalog data at
 # build time.
-localize-messages = ["dep:phf", "dep:fish-gettext-maps"]
+localize-messages = ["dep:fish-gettext"]
 # This feature is used to enable extracting messages from the source code for localization.
 # It only needs to be enabled if updating these messages (and the corresponding PO files) is
 # desired. This happens when running tests via `build_tools/check.sh` and when calling
